requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
pytz==2019.1
PyYAML==5.1
requests==2.21.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
